Add conditionals for debugging experiment scripts.

+++ experiments/setUpContainers.sh
@@ -2,6 +2,13 @@
 date
 set -x
 
+# To avoid execution without passing environment variables
+if [[ (-z "$ID_RSA_PATH") || (-z "$CONTAINER_TAG") || (-z "$CONTAINER_MEM_LIM") ]];
+then
+  echo -e "${RED}THERE ARE VARIABLES MISSING: setUpContainers.sh${NC}"
+  exit
+fi
+
 LAST_IP_NUMBER="198 218 231 242"
 COMMAND="sudo docker run --network=host --cpus=1.0 --memory=${CONTAINER_MEM_LIM} --rm --name container-${CONTAINER_TAG} image-${CONTAINER_TAG} &"
 for i in ${LAST_IP_NUMBER}
